sm8250-common: Switch to AIDL android.hardware.memtrack As we were getting those errors with HIDL android.hardware.memtrack:
D SystemServerTiming: MemtrackProxyService
I hwservicemanager: Since android.hardware.memtrack@1.0::IMemtrack/default is not registered, trying to start it as a lazy HAL (if it's not configured to be a lazy HAL, it may be stuck starting or still starting).
W libc    : Unable to set property "ctl.interface_start" to "android.hardware.memtrack@1.0::IMemtrack/default": PROP_ERROR_HANDLE_CONTROL_MESSAGE (0x20)
I hwservicemanager: Tried to start android.hardware.memtrack@1.0::IMemtrack/default as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.
I hwservicemanager: Since android.hardware.memtrack@1.0::IMemtrack/default is not registered, trying to start it as a lazy HAL (if it's not configured to be a lazy HAL, it may be stuck starting or still starting).
I HidlServiceManagement: getService: Trying again for android.hardware.memtrack@1.0::IMemtrack/default...
I SystemConfig: Reading permissions from /vendor/etc/permissions/android.hardware.bluetooth.xml
I hwservicemanager: Tried to start android.hardware.memtrack@1.0::IMemtrack/default as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.